requirements.txt
discord.py==2.6.3
typing-extensions>=4.12.2
aiofiles==23.2.1
asyncio-throttle==1.0.2
python-dotenv==1.1.1
